+++ packages/widgetbook/lib/src/next/args/const_arg.dart
@@ -2,18 +2,14 @@ import '../../fields/fields.dart';
 import 'arg.dart';
 
 class ConstArg<T> extends Arg<T> {
-  ConstArg(T value) : super(value: value);
+  ConstArg(super.value);
 
   @override
   List<Field> get fields => [];
 
   @override
-  ConstArg<T> init({required String name}) {
-    return ConstArg(value);
-  }
+  ConstArg<T> init({required String name}) => this;
 
   @override
-  T valueFromQueryGroup(Map<String, String> group) {
-    return value;
-  }
+  T valueFromQueryGroup(Map<String, String> group) => value;
 }