+++ packages/widgetbook/lib/next.dart
@@ -22,6 +22,7 @@ export 'src/next/args/arg.dart';
 export 'src/next/args/bool_arg.dart';
 export 'src/next/args/builder_arg.dart';
 export 'src/next/args/color_arg.dart';
+export 'src/next/args/const_arg.dart';
 export 'src/next/args/double_arg.dart';
 export 'src/next/args/duration_arg.dart';
 export 'src/next/args/enum_arg.dart';