+++ packages/widgetbook_generator/lib/src/next/extensions.dart
@@ -1,6 +1,22 @@
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:code_builder/code_builder.dart';
+
+extension ExpressionX on Expression {
+  Expression maybeProperty(
+    String name, {
+    required bool nullSafe,
+  }) {
+    return nullSafe ? this.nullSafeProperty(name) : this.property(name);
+  }
+}
+
+extension ParameterElementX on ParameterElement {
+  bool get requiresArg {
+    return !type.isPrimitive && !type.isNullable && !hasDefaultValue;
+  }
+}
 
 extension DartTypeX on DartType {
   static final typesMeta = {
@@ -39,6 +55,10 @@ extension DartTypeX on DartType {
 
   bool get isPrimitive {
     return isEnum || typesMeta.containsKey(displayName);
+  }
+
+  bool get isNullable {
+    return nullabilitySuffix == NullabilitySuffix.question;
   }
 
   bool get isEnum {