+++ packages/widgetbook/CHANGELOG.md
@@ -6,6 +6,7 @@
 - **EXPERIMENTAL**: Change `Arg.value` to be positional parameter instead of named one. ([#1077](https://github.com/widgetbook/widgetbook/pull/1077))
 - **EXPERIMENTAL**: Allow overriding `Arg.name`. ([#1078](https://github.com/widgetbook/widgetbook/pull/1078))
 - **EXPERIMENTAL**: Support custom args via `MetaWithArgs`. ([#1080](https://github.com/widgetbook/widgetbook/pull/1080))
+- **EXPERIMENTAL**: Add `args` parameter to `Story.setup`. ([#1081](https://github.com/widgetbook/widgetbook/pull/1081))
 
 ## 3.6.0