+++ packages/widgetbook/CHANGELOG.md
@@ -9,6 +9,7 @@
 - **EXPERIMENTAL**: Add `args` parameter to `Story.setup`. ([#1081](https://github.com/widgetbook/widgetbook/pull/1081))
 - **EXPERIMENTAL**: Use `setup` in `Scenario`s. ([#1082](https://github.com/widgetbook/widgetbook/pull/1082))
 - **EXPERIMENTAL**: Add `LeafComponent`. ([#1083](https://github.com/widgetbook/widgetbook/pull/1083))
+- **EXPERIMENTAL**: Support optional parameters arg-generation. ([#1084](https://github.com/widgetbook/widgetbook/pull/1084))
 
 ## 3.6.0